requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
scikit-learn==1.5.0
numpy==1.26.4
httpx==0.27.0
openai==1.35.0        # Used for OpenAI-compatible API (can swap to local)
python-multipart==0.0.9